started styling

+++ src/scripts/views/challengeView.js
@@ -6,13 +6,6 @@ import User from '../models/userModel'
 
 
 
-
-
-
-var currentDate = new Date();
-
-console.log(currentDate)
-
 const ChallengeView = React.createClass({
 
 
@@ -24,6 +17,8 @@ const ChallengeView = React.createClass({
 				STORE._getData()
 			)
 		})
+		this._setUsersLikes()
+		this._setActiveStatus()
 	},
 
 	componentWillUnmount: function() {
@@ -34,13 +29,36 @@ const ChallengeView = React.createClass({
 		return STORE._getData()
 	},
 
+	_setUsersLikes: function() { 
+		if (User.getCurrentUser()) {
+			var u = User.getCurrentUser()
+			var votes = JSON.parse(u.get('tracksLiked'))
+			STORE._set({
+				userVotes: votes
+			})
+		}
+	},
+
+	_setActiveStatus: function() {
+		if(this.state.currentChallenge.attributes.isActive) {
+			STORE._set({
+				activeStatus: true
+			})
+		}
+		else {
+			STORE._set({
+				activeStatus: false,
+			})
+		}
+	},
+
+
 	render: function() {
-		console.log(this.props)
 		return (
 			<div className='challenge-page'>
 				<Header />
 				<ChallengeDetails model={this.state.currentChallenge} /> 
-				<Submit visible={this.state.submissionWindowVisible} id={this.props.id}/>
+				<Submit visible={this.state.submissionWindowVisible} id={this.props.id} />
 				<Tracks collection={this.state.currentTracks} e />
 
 			</div>
@@ -51,6 +69,7 @@ const ChallengeView = React.createClass({
 
 const ChallengeDetails = React.createClass({
 
+
 	_toggleSubmit: function() {
 		
 			STORE._set({
@@ -59,15 +78,56 @@ const ChallengeDetails = React.createClass({
 
 	},
 
+	_winCheck: function() {
+		var deadline = this.props.model.attributes.deadline
+		var today = new Date().toISOString();
+		console.log(deadline, today)
+		if(deadline > today) {
+			return 
+		}
+		else{
+			return
+		}
+
+	},
+
+	_activeCheck: function() {
+		if(this.props.model.attributes.isActive){
+			return true
+		}
+		else {
+
+			return false
+		}
+
+	},
+
+	_dateTranslate: function(isoString) {
+		var dt = isoString.split(/[: T-]/).map(parseFloat);
+     	var d = new Date(dt[0], dt[1] - 1, dt[2], dt[3] || 0, dt[4] || 0, dt[5] || 0, 0);
+    	var dateOut = (d.getMonth()+1) + '/' + d.getDate() + '/' + d.getFullYear() 
+    	console.log(dateOut, d)
+     	return dateOut
+
+	},
+
+
+
 
 	render: function() {
+		//this._winCheck()
 		var challenge = this.props.model
-		console.log(challenge)
 		return (
 			<div className='details-container'>
 				<h1> {challenge.get('title')} </h1>
 				<p> {challenge.get('details')} </p>
-				<button onClick={this._toggleSubmit}> Submit a Track! </button> 
+
+				{this._activeCheck() ? 
+						(<div> <h3> deadline: {this._dateTranslate(challenge.attributes.deadline)} </h3> <button id="mycrazybutt" onClick={this._toggleSubmit}> Submit a Track! </button> </div>)
+
+						 : (<h3> won </h3>)}
+
+				 
 			</div>
 			)
 	}
@@ -88,21 +148,28 @@ const Submit = React.createClass({
 		_submissionsCheck: function(eventObj) {
 			eventObj.preventDefault()
 
-			var ChallId = this.props.id
-			var subs = JSON.parse(User.getCurrentUser().get('submissions'))
-			if (ChallId in subs) {				
-				var quest = confirm("You can only submit one track per challenge. Do you wish to overwrite previously submmitted track? ")
-				if (quest === true) {
-					var track = subs[ChallId]
-					ACTIONS.deleteTrack(track)
-					this._submitTrack(eventObj)
+			if (User.getCurrentUser()){
+
+				var ChallId = this.props.id
+				var subs = JSON.parse(User.getCurrentUser().getCurrentUser().get('submissions'))
+				if (ChallId in subs) {				
+					var quest = confirm("You can only submit one track per challenge. Do you wish to overwrite previously submmitted track? ")
+					if (quest === true) {
+						var track = subs[ChallId]
+						ACTIONS.deleteTrack(track)
+						this._submitTrack(eventObj)
+					}
+					else {
+						this._toggleSubmit()
+					}
 				}
 				else {
-					this._toggleSubmit()
+					this._submitTrack(eventObj)
 				}
 			}
 			else {
-				this._submitTrack(eventObj)
+				alert('you must have an account to post a track')
+				this._toggleSubmit()
 			}
 		},
 
@@ -127,7 +194,6 @@ const Submit = React.createClass({
 			//})
 		},
 
-	
 
 		render: function() {
 
@@ -137,12 +203,11 @@ const Submit = React.createClass({
 		    }
 
 
-			console.log(User.getCurrentUser())
 
 			return (
 				
 				<form className='submission-container' style={displayStyle} onSubmit={this._submissionsCheck}>
-					<h2>Post your stuff</h2>
+					<h2>Post your track</h2>
 					<input type="title" 	   name="title" 	   placeholder="title" />
 					<input type="text" 	       name="URL" 		   placeholder="track link" />
 					<input type="description"  name="description"  placeholder="description (optional)" />
@@ -169,29 +234,66 @@ const Tracks = React.createClass({
 
 const Track = React.createClass({
 
-	_vote4: function(e) {
-		if(!User) {
+	_vote4: function() {
+		var currentTrack = this.props.model.attributes
+
+
+		if(!User.getCurrentUser()) {
 			alert('please login to vote')
 		}
-		else if (User) {
-			ACTIONS.saveVote
+		else if (User.getCurrentUser()) {
+
+			var userLikes = JSON.parse(User.getCurrentUser().get('tracksLiked'))
+				if (currentTrack._id === userLikes[currentTrack.challengeId]){
+					ACTIONS.saveVote(currentTrack.challengeId," ")
+				}
+
+				else {
+				ACTIONS.saveVote(currentTrack.challengeId,currentTrack._id)
+				}	
 
 		}
 
 	},
 
-	_checkCheck: function(e) {
+
+
+	//renders either a filled in or outlined heart
+	_voteCheck: function() {
+		if (User.getCurrentUser()) {
+			var currentTrack = this.props.model.attributes
+			var u = User.getCurrentUser()
+			var votes = JSON.parse(u.get('tracksLiked'))
+			if(currentTrack._id === votes[currentTrack.challengeId]){
+				return true
+			}
+			else {
+				return false
+			}
+		}
+		else {
+			return false
+		}
 	},
 
-	render: function() {
 
+
+
+	//<i class="fa fa-star" aria-hidden="true"></i>    <i className="fa fa-star-o" aria-hidden="true"></i>
+	render: function() {
+		var currentTrack = this.props.model.attributes
+		console.log(currentTrack)
 
 		return (
 			<div className='track'> 
-				<h2> {this.props.model.attributes.title} </h2>
-				<a href={this.props.model.attributes.link}> check it out! </a>
-				<div className="vote-box"> 
-					<i className="fa fa-star-o" aria-hidden="true"></i>
+				<h2> {currentTrack.title} </h2>
+				<a href={currentTrack.link}> check it out! </a>
+				<div className="vote-box" onClick={this._vote4}> 
+					{this._voteCheck() ? (
+						<i className="fa fa-star" aria-hidden="true"></i>
+						) : (
+						<i className="fa fa-star-o" aria-hidden="true"></i>
+						)}
 				</div>
 			</div>
 			)
